Feature/calendar

+++ src/main/java/kr/carz/savecar/dto/MorenDTO.java
@@ -25,7 +25,7 @@ public class MorenDTO {
     private String carOld;
     private String carEngine;
     private String carAttribute01;
-    private Float carPrice;
+    private double carPrice;
     private String orderEnd;
     private Long rentIdx;
     private List<String> carImageList;
